add plugin for environment variables support. add password protect to delete route

+++ src/main/java/com/covid19trackerv2/controller/CountryController.java
@@ -4,6 +4,7 @@ import com.covid19trackerv2.model.country.CountryDoc;
 import com.covid19trackerv2.model.state.StateDoc;
 import com.covid19trackerv2.repository.CountryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.ResponseEntity;
@@ -11,12 +12,16 @@ import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @RestController
 @CrossOrigin
 @RequestMapping("/api/country")
 public class CountryController {
+
+    @Autowired
+    Environment environment;
 
     @Autowired
     private CountryRepository countryRepo;
@@ -56,7 +61,9 @@ public class CountryController {
 
 
     @DeleteMapping("delete_countries")
-    public void deleteAllCountries() {
-        this.countryRepo.deleteAll();
+    public void deleteAllCountries(@RequestBody Map<String, String> password) {
+        if(password.get("password").equals(environment.getProperty("DB_PASSWORD"))) {
+            this.countryRepo.deleteAll();
+        }
     }
 }
